refactor: misc clean up
remove requirements.txt

clean up .gitignore

move init_database test into test_init_database.py
