+++ .gitignore
@@ -1,15 +1,3 @@
-# Generated files
-lean/
-lean_files/
-lean_data
-src/lean_data
-src/lean_data.old
-output.json
-llm_proofs.json
-
-# logs
-*.log
-
 # Python
 __pycache__/
 *.pyc
@@ -18,7 +6,6 @@ __pycache__/
 .env
 .venv
 env/
-venv/
 ENV/ 
 
 # Lean .lake dirs
