--- requirements.txt
@@ -1,1 +1,0 @@
-requests>=2.31.0
